+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidyselect (development version)
 
 * `matches()` now uses `perl = TRUE` by default. This makes it more consitent with regular expressions in stringr (#330).
+* `eval_select()` now fails when data has duplicate names and a character vector is provided as input (#346).
 
 * `eval_select()` and `eval_relocate()` gain a new `error_arg` argument that can be specified to throw a better error message when `allow_empty = FALSE`  or `allow_rename = FALSE` (@olivroy, #327).